+++ src/cleaning.html
@@ -82,100 +82,3 @@
     <script src="main-cleaning.js" type="module"></script>
   </body> 
 </html>
-
-          <!-- Card base template -->
-          <!-- <li class="card-container" itemscope itemtype="https://schema.org" itemprop="pant">-->
-            <!-- Front card funciona -->
-            <!-- <article class="front-card" id="front-card">
-              <h2>Agave Azul</h2>
-              <div class="top-card">
-                <img alt="Agave Azul" src="resources/Images/agave-azul.png">
-                <dl itemscope itemtype="https://schema.org" class="facts">
-                  <div class="amounts">
-                    <dt class="amount">Agua</dt>
-                    <dd>
-                      <img alt="Gota" id="minimum-water" src="resources/Icons/agua-activa.png">
-                      <img alt="Gota" id="medium-water" src="resources/Icons/agua-activa.png">
-                      <img alt="Gota" id="maximum-water" src="resources/Icons/agua-inactiva.png" class="inactive">
-                    </dd>
-                  </div>
-                  <div class="amounts">
-                    <dt class="amount">Luz</dt>
-                    <dd>
-                      <img alt="Sol" id="minimum-light" src="resources/Icons/luz-activa.png">
-                      <img alt="Sol" id="medium-light" src="resources/Icons/luz-inactiva.png" class="inactive">
-                      <img alt="Sol" id="maximum-light" src="resources/Icons/luz-inactiva.png" class="inactive">
-                    </dd>
-                  </div>
-                  <div class="amounts">
-                    <dt class="amount">Cuidado</dt>
-                    <dd>
-                      <img alt="Semaforo" id="minimum-care" src="resources/Icons/cuidado-activa.png">
-                      <img alt="Semaforo" id="medium-care" src="resources/Icons/cuidado-activa.png">
-                      <img alt="Semaforo" id="maximum-care" src="resources/Icons/cuidado-inactiva.png" class="inactive">
-                    </dd>
-                  </div>
-                </dl>
-              </div>
-              <div class="botom-card">
-                <p>Planta suculenta usada en la producción de tequila.</p>
-                <button class="details-button" id="details-button">Detalles</button>
-              </div>
-            </article> -->
-            <!-- Back card, top funciona, icons y botom card-->
-            <!-- <article id="back-card" class="hide">
-              <h2>Agave Azul</h2>
-              <div class="top-card">
-                <img alt="Plant Name" src="resources/Images/agave-azul.png">
-                <dl itemscope itemtype="https://schema.org">
-                  <dt class="scientific-name">Agave tequilana</dt>
-                  <dt>Familia botánica</dt><dd>Asparagaceae</dd>
-                  <dt>Lugar de origen</dt><dd>México</dd>
-                </dl>
-              </div>
-              <div class="bottom-card">
-                <dl id="description">
-                  <dt>Usos</dt><dd>Producción de tequila, jardinería ornamental.</dd>
-                  <dt>Datos climáticos</dt><dd>Prefiere climas secos y suelos bien drenados.</dd>
-                  <dt>Mantenimiento</dt><dd>Riego limitado, resistente a la sequía y al calor.</dd>
-                </dl>
-              </div>
-              <div id="icons">
-                <img class="modal-statistics-button" alt="Estadísticas" src="resources/Icons/estadisticas.png">
-                <img class="modal-description-button" id="modal-description-button" alt="Descripción" src="resources/Icons/descripcion.png">
-                <input type="image" id="flip-button" class="flip-button" alt="Regresar" src="resources/Icons/Regresar.png">
-              </div>
-            </article>
-          </li> -->
-          <!-- Modals -->
-          <!-- <dialog class="modal-boxes">
-            <article id="description-modal" class="description-modal"> 
-              <h3>Plant Name</h3>
-              <p>Description</p>
-            </article>
-            <article id="statistics-modal" class="statistics-modal"> 
-              <h3>Category</h3>
-              <div class="statistics">
-                <div class="graph-legend">
-                  <div class="legend">
-                    <div class="water-square"></div>
-                    <p>Agua</p>
-                  </div>
-                  <div class="legend">
-                    <div class="light-square"></div>
-                    <p>Luz</p>
-                  </div>
-                  <div class="legend">
-                    <div class="care-square"></div>
-                    <p>Cuidado</p>
-                  </div>
-                </div>
-                <div class="statistics-totals">
-                  <h5 class="water">Mucha</h5>
-                  <h5 class="light">Regular</h5>
-                  <h5 class="care">Poca</h5>
-                </div>
-              </div>
-              <h4>Qué cuidados necesita este tipo de plantas?</h4>
-            </article> 
-          </dialog>  -->
